Removed BLASVector + moved vector operations to VectorView + delegated Vector operations to underlying VectorView
